refactor: initializer 클래스 app으로 이동
