+++ app/src/main/java/com/techcourse/HandlerAdaptersInitializer.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet;
+package com.techcourse;
 
 import com.interface21.webmvc.servlet.mvc.tobe.adapter.AnnotationHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.adapter.HandlerAdapters;
